Add files via upload
This project was a major project for me. This is the famous Microsoft Adventure Works project where the Product, Categories and the Colors tables had to be merged together and create a fact relationship out of the product table. I started by renaming the tables columns and creating a delimiter by the dash to separate the category code, category size and category color. I also made a pivot table out of this analysis so check it out and let me know what you think but first make sure to download it and enable edit and then go into the data modeling section to see all the relationships.